docs: editorial for recent documentation updates (#4395)
Relevant to these PRs:

- #4380
- #4381
- #4382
- #4383
- #4391

In general this content was excellent - thanks @sarahxsanders! - but I
wanted to add a few minor updates. I've included comments in a few
places with reasoning.

If this PR is too much to review all at once, I'm happy to break it up
as I did review of each PR separately.